refactor cache store classes
