+++ imagetool/src/main/java/com/oracle/weblogic/imagetool/cachestore/CacheStoreFactory.java
@@ -1,13 +1,12 @@
 // Copyright (c) 2019, Oracle Corporation and/or its affiliates.  All rights reserved.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package com.oracle.weblogic.imagetool.cachestore.meta;
+package com.oracle.weblogic.imagetool.cachestore;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
 
-import com.oracle.weblogic.imagetool.api.meta.CacheStore;
 import com.oracle.weblogic.imagetool.util.Constants;
 
 public class CacheStoreFactory implements Supplier<CacheStore> {
