+++ imagetool/src/test/java/com/oracle/weblogic/imagetool/cachestore/meta/FileCacheStoreTest.java
@@ -3,7 +3,8 @@
 
 package com.oracle.weblogic.imagetool.cachestore.meta;
 
-import com.oracle.weblogic.imagetool.api.meta.CacheStore;
+import com.oracle.weblogic.imagetool.cachestore.CacheStore;
+import com.oracle.weblogic.imagetool.cachestore.CacheStoreFactory;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
